backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
bcrypt==4.1.2
Pillow==10.2.0
PyPDF2==3.0.1
python-magic==0.4.27
Werkzeug==3.0.1
SQLAlchemy==2.0.25
python-dotenv==1.0.0
cryptography==42.0.0
requests==2.31.0
user-agents==2.2.0
gunicorn==21.2.0